Add support for 'theme packs'
Right now only replacing the icon in the action bar is supported.
That's why the term "Icon Pack" is used in the UI.

Known bug: In a PreferenceActivity the second level still shows the
default app icon in the action bar :(

Icon packs need to include an activity with an intent filter for
the action "org.k9mail.THEME_PACK".

Example:

<activity android:name=".SomeActivity">
  <intent-filter>
    <action android:name="org.k9mail.THEME_PACK" />
  </intent-filter>

  <!-- Version number for the theme pack format -->
  <meta-data android:name="version" android:value="1"/>

  <meta-data android:name="name" android:value="Fancy icon"/>
  <meta-data android:name="author" android:value="cketti"/>
  <meta-data android:name="icon_app"
             android:resource="@drawable/ic_app"/>
</activity>